mcp_servers/arada-sql/requirements.txt
# Arada SQL MCP Server dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pandas>=2.0.0
